requirements.txt
langchain>=0.1.16
duckduckgo-search>=4.3
beautifulsoup4>=4.12
requests>=2.31
kaggle>=1.6
huggingface_hub>=0.23
python-dotenv>=1.0
groqcloud>=0.4
streamlit>=1.35
